Bug #1425097 misc: sprdwcn: pclint warning eliminate
[root cause  ] date type conversion error
[changes     ] sprdwcn
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:SC7731E_ANDROID11_TRUNK
[test case]:pclint tool
[reviewers   ] yaoguang.chen songhe.wei
[change_type ] feature_bugfix
[tag_product ] common

This is the pclint warning,and exit data type conversion error,so
need to forced conversion
Test: pclint scanning

+++ drivers/misc/sprdwcn/platform/bufring.c
@@ -141,7 +141,8 @@ int mdbg_ring_read(struct mdbg_ring_t *ring, void *buf, int len)
 	}
 	MDBG_RING_LOCK(ring);
 	cont_len = mdbg_ring_readable_len(ring);
-	read_len = (unsigned int)(cont_len >= len ? len : cont_len);
+	read_len = (unsigned int)(cont_len >= len ? len :
+				  (unsigned int)cont_len);
 	pstart = mdbg_ring_start(ring);
 	pend = mdbg_ring_end(ring);
 	WCN_LOG("read_len=%d", read_len);
